requirements.txt
requests
beautifulsoup4
discord.py
python-dotenv
Pillow
